scheduler/requirements.txt
azure-functions==1.20.0
httpx==0.27.0
